backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.0
attrs==20.3.0
black==20.8b1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
coverage==5.3
distlib==0.3.1
Django==3.1.3
django-filter==2.4.0
django-oauth-toolkit==1.3.3
djangorestframework==3.12.2
filelock==3.0.12
flake8==3.8.4
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
Markdown==3.3.3
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.4
pathspec==0.8.0
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pynt==0.8.2
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytz==2020.4
regex==2020.10.28
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tox==3.20.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.1.0
zipp==3.4.0
